requirements.txt
python-dotenv
azure-ai-documentintelligence
anthropic
openai
httpx
PyMuPDF
